# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API для управления задачами и командами",
        "title": "SquadSync API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает задачи с фильтрацией и пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Список задач",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID команды",
                        "name": "team_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус (todo, in_progress, done)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ID исполнителя",
                        "name": "assignee_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagTaskListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает задачу в указанной команде",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создание задачи",
                "parameters": [
                    {
                        "description": "Данные задачи",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Частичное обновление полей задачи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Обновление задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновляемые поля",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.UpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает лог изменений полей задачи",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "История изменений задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagHistoryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_adapter_http_handler.swagErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_adapter_http_handler.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_http_handler.HistoryEntry": {
            "type": "object",
            "properties": {
                "changed_by": {
                    "type": "integer"
                },
                "field_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "new_value": {
                    "type": "string"
                },
                "old_value": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_http_handler.TaskListResponse": {
            "type": "object",
            "properties": {
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_adapter_http_handler.TaskResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_adapter_http_handler.TaskResponse": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "integer"
                },
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "team_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_http_handler.UpdateTaskRequest": {
            "type": "object",
            "properties": {
                "assignee_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/task_vault_internal_domain.Status"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_http_handler.swagErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_adapter_http_handler.swagHistoryListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_adapter_http_handler.HistoryEntry"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "internal_adapter_http_handler.swagTaskListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_adapter_http_handler.TaskListResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_adapter_http_handler.swagTaskResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_adapter_http_handler.TaskResponse"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "task_vault_internal_domain.Status": {
            "type": "string",
            "enum": [
                "todo",
                "in_progress",
                "done"
            ],
            "x-enum-varnames": [
                "StatusTodo",
                "StatusInProgress",
                "StatusDone"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT-токен (формат: \"Bearer \u003ctoken\u003e\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}